+++ packages/d2bd/src/unsafe_local_helper.rs
@@ -1385,6 +1385,8 @@ mod tests {
             item_id: ProtocolToken::parse("browser").unwrap(),
             argv: ConfiguredArgv::new(vec![arg.to_owned()]).unwrap(),
             graphical: false,
+            realm_accent_color: d2b_contracts::unsafe_local_wire::RealmAccentColor::new("#336699")
+                .unwrap(),
         }
     }